Moved editor code to Editor folder to avoid build system, preserved references
