+++ VR Maze/Assets/Scripts/ReadOnlyAttribute.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEditor;
 
 namespace Assets.EditorTools
 {
@@ -9,23 +8,4 @@ namespace Assets.EditorTools
     /// Makes a public property readonly in the Unity editor
     /// </summary>
     public class ReadOnlyAttribute : PropertyAttribute { };
-
-    /// <summary>
-    /// Unity Editor property drawer for ReadOnly properties
-    /// </summary>
-    [CustomPropertyDrawer(typeof(ReadOnlyAttribute))]
-    public class ReadOnlyDrawer : PropertyDrawer
-    {
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            return EditorGUI.GetPropertyHeight(property, label, true);
-        }
-
-        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
-        {
-            GUI.enabled = false;
-            EditorGUI.PropertyField(position, property, label, true);
-            GUI.enabled = true;
-        }
-    }
 }
